# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Chamber Commissions API — سامانه کمیسیون‌های تخصصی
  version: 1.0.0
  description: 'REST API of the commissions platform. All business endpoints require `Authorization: Bearer <accessToken>`.


    Real-time: Socket.IO at `/socket.io` (auth: `{ token }`). Emit `meeting:join` with a meeting id, then receive `meeting.updated`, `attendance.updated`
    (officers), `quorum.updated`, `agenda.updated`, `comment.created`, `vote.opened`, `vote.progress`, `vote.closed`, and personal `notification`
    events.


    Errors: `{ error: { code, message, details? } }` with Persian messages.'
servers:
- url: http://localhost:4000
security:
- bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: &id001
              type: string
            message: *id001
            details:
              type: array
              items:
                type: object
                properties:
                  path: *id001
                  message: *id001
...
    ResolutionCreate:
      type: object
      required:
      - commissionId
      - text
      properties:
        commissionId: *id004
        meetingId: *id004
        agendaItemId: *id004
        voteSessionId: *id004
        text: *id001
        ownerId: *id004
        addressee: *id001
        dueDate: &id006
          type: string
          format: date
          example: '2026-10-20'
        priority:
          type: string
          enum:
          - low
          - normal
          - high
          - urgent
        kpi: *id001
paths:
...
  /api/resolutions:
    post:
      tags:
      - Resolutions
      summary: ایجاد مصوبه (+Task)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolutionCreate'
    get:
      tags:
      - Resolutions
      summary: فهرست مصوبات (?commissionId&status&mine&overdue&q)
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/resolutions/{id}:
    get:
      tags:
      - Resolutions
      summary: جزئیات مصوبه، تاریخچه و مستندات
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
    patch:
      tags:
      - Resolutions
      summary: ویرایش/لغو مصوبه
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: *id001
                ownerId: *id004
                dueDate: *id006
                status: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
  /api/resolutions/{id}/progress:
    patch:
      tags:
      - Resolutions
      summary: ثبت پیشرفت توسط مسئول اجرا
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                progress: *id002
                note: *id001
                documentId: *id004
                submit: *id003
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
  /api/resolutions/{id}/review:
    post:
      tags:
      - Resolutions
      summary: تأیید یا برگشت نتیجه توسط دبیر
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approve: *id003
                note: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
  /api/tasks:
    get:
      tags:
      - Resolutions
      summary: وظایف باز من
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/commissions/{id}/issues:
    get:
      tags:
      - Issues
      summary: مسائل کمیسیون
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
    post:
      tags:
      - Issues
      summary: ثبت مسئله
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: *id001
                description: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
  /api/issues/{id}:
    patch:
      tags:
      - Issues
      summary: ویرایش مسئله
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
  /api/referrals:
    post:
      tags:
      - Issues
      summary: ارجاع کارشناسی
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                commissionId: *id004
                issueId: *id004
                expertId: *id004
                request: *id001
                dueDate: *id006
    get:
      tags:
      - Issues
      summary: 'ارجاعات (کارشناس: ارجاعات خود؛ ?commissionId برای دبیر)'
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/referrals/{id}/answer:
    post:
      tags:
      - Issues
      summary: پاسخ کارشناس
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                response: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
...